+++ opentok/opentok.py
@@ -32,7 +32,7 @@ from .stream import Stream
 from .streamlist import StreamList
 from .sip_call import SipCall
 from .broadcast import Broadcast, BroadcastStreamModes
-from .websocket_audio_connection import WebsocketAudioConnection
+from .websocket_audio_connection import WebSocketAudioConnection
 from .exceptions import (
     ArchiveStreamModeError,
     BroadcastHLSOptionsError,
@@ -49,7 +49,7 @@ from .exceptions import (
     SetStreamClassError,
     BroadcastError,
     DTMFError,
-    InvalidWebsocketOptionsError,
+    InvalidWebSocketOptionsError,
     InvalidMediaModeError  
 )
 
@@ -1804,14 +1804,14 @@ class Client(object):
         else:
             raise RequestError("An unexpected error occurred", response.status_code)
 
-    def stream_audio_to_websocket(self, session_id: str, opentok_token: str, websocket_options: dict):
+    def connect_audio_to_websocket(self, session_id: str, opentok_token: str, websocket_options: dict):
         """
-        Connects audio streams to a specified websocket URI.
-        For more information, see the `Audio Streamer developer guide <https://tokbox.com/developer/guides/audio-streamer/>`.
+        Connects audio streams to a specified WebSocket URI.
+        For more information, see the `Audio Connector developer guide <https://tokbox.com/developer/guides/audio-streamer/>`.
 
-        :param String 'session_id': The OpenTok session ID that includes the OpenTok streams you want to include in the WebSocket stream. The Audio Streamer feature is only supported in routed sessions (sessions that use the OpenTok Media Router).
-        :param String 'opentok_token': The OpenTok token to be used for the Audio Streamer connection to the OpenTok session.
-        :param Dictionary 'websocket_options': Included options for the websocket.
+        :param String 'session_id': The OpenTok session ID that includes the OpenTok streams you want to include in the WebSocket stream. The Audio Connector feature is only supported in routed sessions (sessions that use the OpenTok Media Router).
+        :param String 'opentok_token': The OpenTok token to be used for the Audio Connector connection to the OpenTok session.
+        :param Dictionary 'websocket_options': Included options for the WebSocket.
             String 'uri': A publicly reachable WebSocket URI to be used for the destination of the audio stream (such as "wss://example.com/ws-endpoint").
             List 'streams' Optional: A list of stream IDs for the OpenTok streams you want to include in the WebSocket audio. If you omit this property, all streams in the session will be included.
             Dictionary 'headers' Optional: An object of key-value pairs of headers to be sent to your WebSocket server with each message, with a maximum length of 512 bytes.
@@ -1826,14 +1826,14 @@ class Client(object):
 
         logger.debug(
             "POST to %r with params %r, headers %r, proxies %r",
-            self.endpoints.get_audio_streamer_url(),
+            self.endpoints.get_audio_connector_url(),
             json.dumps(payload),
             self.get_json_headers(),
             self.proxies,
         )
 
         response = requests.post(
-            self.endpoints.get_audio_streamer_url(),
+            self.endpoints.get_audio_connector_url(),
             json=payload,
             headers=self.get_json_headers(),
             proxies=self.proxies,
@@ -1841,7 +1841,7 @@ class Client(object):
         )
 
         if response and response.status_code == 200:
-            return WebsocketAudioConnection(response.json())
+            return WebSocketAudioConnection(response.json())
         elif response.status_code == 400:
             """
             The HTTP response has a 400 status code in the following cases:
@@ -1852,15 +1852,15 @@ class Client(object):
         elif response.status_code == 403:
             raise AuthError("You passed in an invalid OpenTok API key or JWT token.")
         elif response.status_code == 409:
-            raise InvalidMediaModeError("Only routed sessions are allowed to initiate Audio Streamer WebSocket connections.")
+            raise InvalidMediaModeError("Only routed sessions are allowed to initiate Audio Connector WebSocket connections.")
         else:
             raise RequestError("An unexpected error occurred", response.status_code)
 
     def validate_websocket_options(self, options):
         if type(options) is not dict:
-            raise InvalidWebsocketOptionsError('Must pass websocket options as a dictionary.')
+            raise InvalidWebSocketOptionsError('Must pass WebSocket options as a dictionary.')
         if 'uri' not in options:
-            raise InvalidWebsocketOptionsError('Provide a websocket URI.')
+            raise InvalidWebSocketOptionsError('Provide a WebSocket URI.')
 
     def _sign_string(self, string, secret):
         return hmac.new(