+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release v3.4.0
 
-- Support for Audio Streamer (lite) API via `stream_audio_to_websocket` method
+- Support for Audio Connector API via `connect_audio_to_websocket` method
 
 # Release v3.3.0